+++ DungeonSeeker/Assets/Scripts/SaveData.cs
@@ -12,10 +12,10 @@ public class SaveData
     public float dashCoolTimePer = 0;
     public float AttackCoolTimePer = 0;
     public float DmgPer = 0;
-    public int PlayerPlat = 0;
+    public float PlayerPlat = 0;
     public bool skillOn = false;
     public float skilldurationPer = 0;
     public float skillCoolTimePer = 0;
-    public int AirJumpCountMaxPer = 0;
-    public int HpPotionMaxPer = 0;
+    public float AirJumpCountMaxPer = 0;
+    public float HpPotionMaxPer = 0;
 }
